refactor(NcVNodes): migrate component to Typescript
