+++ src/components/index.ts
@@ -64,7 +64,6 @@ export { default as NcListItem } from './NcListItem/index.js'
 export { default as NcListItemIcon } from './NcListItemIcon/index.js'
 export { default as NcLoadingIcon } from './NcLoadingIcon/index.js'
 export { default as NcModal } from './NcModal/index.js'
-export { default as NcVNodes } from './NcVNodes/index.js'
 export { default as NcNoteCard } from './NcNoteCard/index.js'
 export { default as NcPasswordField } from './NcPasswordField/index.js'
 export { default as NcPopover } from './NcPopover/index.js'
@@ -83,3 +82,4 @@ export { default as NcTextField } from './NcTextField/index.js'
 export { default as NcTimezonePicker } from './NcTimezonePicker/index.ts'
 export { default as NcUserBubble } from './NcUserBubble/index.js'
 export { default as NcUserStatusIcon } from './NcUserStatusIcon/index.js'
+export { default as NcVNodes } from './NcVNodes/index.ts'
